Save the last checkpoint in the PAX training loop.
Currently, the last checkpoint of the training is not saved if the total number of training steps is not fully divisible by the checkpoint saving interval. This CL enables saving the last checkpoint in such scenarios.

A list of the changes:
- Save the checkpoint before checking the for-loop exit condition.
- Change the loop exit condition to ">=".
- Add an option to disable the step_id check in the checkpoint manager.

The first two changes are necessary to avoid subtracting model_states.step by 1 for the last checkpoint, which is non-trivial for GDA/PJIT model states.

PiperOrigin-RevId: 470761215