Trace step functions per eval and decode inputs, allowing multiple batch sizes.
Previously, in `train + {eval,decode}`, if the global batch size of the train batch size was not equal to that of the eval batch size (which was an implicit requirement) the `_SpmdModelPartitioner._get_step_fn` would silently drop examples leading to subtly wrong results.

One subtlety is how we take care of using the right unpadded batch size to unpad, for cases where the per-core batch size < 1, so we explicitly handle two copies of input params. One padded for small bsz and other unpadded.

PiperOrigin-RevId: 490065100