Pax + Fiddle migration: Change `layer_p.cls` to `pax_fiddle.get_callable(layer_p)`.
Also change `layer_p.cls = value` to `pax_fiddle.set_callable(layer_p, value)`.

Currently, layer templates are encoded using `pax_fiddle.Config`, a subclass of `fdl.Config` that provides some backwards-compatibility stubs such as the `cls` property.  But we'd like to remove this subclass, and move to using the standard `fdl.Config` object instead.  This change is a step in that direction.

PiperOrigin-RevId: 495074048