Always use train_inputs_shape_dtype to get the train state metadata.
Sets train_inputs_shape_dtype of the partitioner in all eval loops when
always_use_train_for_model_init is False, this makes sure that
train_inputs_shape_dtype is present when partitioner.get_train_state_metadata()
is called.

This also helps to make sure the partition spec of the train state is fixed
through the lifecycle of the train/eval pipeline, since it's determined by the
var_weight_hparams, which is determined by the input specs and also should be
fixed.

PiperOrigin-RevId: 505216373