[Pax/Fiddle] Rename instantiated Learner._optimizer to remove name conflicts between self._xyz and self.hparams.xyz.
As part of the Fiddle migration, we will be removing `self.hparams` and moving parameters named `self.hparams.xyz` to `self._xyz`. As such, fields which are currently named `self._xyz` and are simply instantiations of the corresponding HParams field are being renamed to `self._xyz_inst`. We plan to follow up after the initial migration to simplify these fields to use a "direct instantiation" pattern where possible, allowing __init__ to simply take an instance constructed as part of fdl.build().

PiperOrigin-RevId: 506134523