[pax+fiddle] Update the `SingleTask.model` field to use optional direct instantiation.
As before, if the `model` argument to the `SingleTask` constructor is a `fdl.Config[BaseModel]`, then it will be instantiated by the constructor.

But now, the `model` argument can also contain a `BaseModel`, in which case it will be used as-is.

After we migrate SingleTask to use Fiddle (rather than HParams) for configuration, we can switch this to always use direct instantiation (i.e., drop the support for having the constructor take a `fdl.Config[BaseModel]`).

PiperOrigin-RevId: 509280256