[pax+fiddle] Update `infer_and_write_pmap` to get the name of each input from the input itself (not its hparams).
Previously, this code relied on the (undocumented) fact that `instantiate(p: LingvoInputAdapter.HParams)` has the side-effect of mutating `p` in place, modifying its `name` attribute.  In particular, if `p.name` is empty, then it is set to `p.input.name`.  This will no longer be the case once we migrate from HParams to Fiddle.

PiperOrigin-RevId: 509517528