Set `summary_verbosity` as an `HParam` for Jax `SingleTask`s
Having a `summary_verbosity` level specified at the task level is pretty useful; it allows users to set the verbosity level of their training/eval tasks at the task level, instead of having to explicitly sprinkle `add_summary()` calls throughout the code manually. In this CL, a `summary_verbosity` is introduced at the task level, which is then plumbed through to various utils in `trainer_lib` where the task is used.

PiperOrigin-RevId: 524304494