Remove hstack for passthrough values in split_out_xla_unsupported_batch.
This previously appeared necessary for a specific pipeline using LingvoInputAdaptor; however, other pipelines (e.g. SeqIO) don't appear to have the extra dimension, making the hstack superflous and causing issues.

Therefore, we change this general implementation to use the value as is. For pipelines which do need the extra concatenation, it can be handled inside the model instead.

PiperOrigin-RevId: 548224329